All Enemy logic updated, adding boss

+++ Mercs/Source/Main.cpp
@@ -25,7 +25,7 @@ int main(int argc, char* argv[])
 	const int FPS = 60;
 	const int frameDelay = 1000 / FPS;
 
-	Uint32 frameStart;
+	Uint64 frameStart;
 	int frameTime;
 
 	int main_return = EXIT_FAILURE;
